Move all source code in src
