+++ src/controller/query/create.js
@@ -6,7 +6,7 @@ var RestError = require('rest-error');
 
 // __Module Definition__
 var decorator = module.exports = function (options, protect) {
-  var baucis = require('../..');
+  var baucis = require('../../..');
   var controller = this;
 
   controller.query('post', function (request, response, next) {
