+++ src/controller/query/update.js
@@ -9,7 +9,7 @@ var validOperators = [ '$set', '$push', '$pull', '$addToSet', '$pop', '$pushAll'
 
 // __Module Definition__
 var decorator = module.exports = function (options, protect) {
-  var baucis = require('../..');
+  var baucis = require('../../..');
   var controller = this;
 
   function checkBadUpdateOperatorPaths (operator, paths) {
