+++ src/controller/request/allow.js
